file-processor/requirements.txt
google-cloud-storage
google-cloud-vision
google-cloud-speech
google-cloud-pubsub
google-cloud-aiplatform
vertexai
cloudevents
python-magic
requests
functions-framework
python-dotenv
PyPDF2
python-docx
pandas
bigframes